refactor: user 관련 패키지 변경 및 코드 정리
