+++ src/main/kotlin/kr/co/shophub/shophub/user/controller/AuthController.kt
@@ -1,14 +1,12 @@
 package kr.co.shophub.shophub.user.controller
 
 import kr.co.shophub.shophub.global.dto.CommonResponse
-import kr.co.shophub.shophub.global.error.UnauthenticatedException
-import kr.co.shophub.shophub.user.controller.dto.reqeust.JoinRequest
-import kr.co.shophub.shophub.user.controller.dto.reqeust.LoginRequest
-import kr.co.shophub.shophub.user.controller.dto.reqeust.TokenRequest
-import kr.co.shophub.shophub.user.controller.dto.response.TokenResponse
-import kr.co.shophub.shophub.user.controller.dto.response.UserResponse
+import kr.co.shophub.shophub.user.dto.JoinRequest
+import kr.co.shophub.shophub.user.dto.LoginRequest
+import kr.co.shophub.shophub.user.dto.TokenRequest
+import kr.co.shophub.shophub.user.dto.TokenResponse
+import kr.co.shophub.shophub.user.dto.UserResponse
 import kr.co.shophub.shophub.user.service.AuthService
-import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -16,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-@RequestMapping("/auth")
+@RequestMapping("/api/v1/auth")
 class AuthController(
     private val authService: AuthService,
 ) {
