+++ src/main/kotlin/kr/co/shophub/shophub/user/model/ProviderType.kt
@@ -1,4 +1,4 @@
-package kr.co.shophub.shophub.user.domain
+package kr.co.shophub.shophub.user.model
 
 enum class ProviderType {
     GOOGLE,
