+++ src/main/kotlin/kr/co/shophub/shophub/user/model/User.kt
@@ -1,4 +1,4 @@
-package kr.co.shophub.shophub.user.domain
+package kr.co.shophub.shophub.user.model
 
 import jakarta.persistence.*
 import org.springframework.security.crypto.password.PasswordEncoder
