+++ src/main/kotlin/kr/co/shophub/shophub/user/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package kr.co.shophub.shophub.user.repository
 
-import kr.co.shophub.shophub.user.domain.User
+import kr.co.shophub.shophub.user.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserRepository : JpaRepository<User, Long> {
